Sequences, removing redundant casts

+++ src/irvine/oeis/a001/A001583.java
@@ -21,7 +21,7 @@ public class A001583 extends Sequence1 {
       if (mP > Integer.MAX_VALUE) {
         throw new UnsupportedOperationException();
       }
-      if (mP % 5 == 1 && Functions.FIBONACCI.z((long) (int) (mP / 5)).mod(mP) == 0) {
+      if (mP % 5 == 1 && Functions.FIBONACCI.z(mP / 5).mod(mP) == 0) {
         return Z.valueOf(mP);
       }
     }
