+++ src/irvine/oeis/a321/A321520.java
@@ -25,7 +25,7 @@ public class A321520 extends GeneralizedEulerTransform {
 
   @Override
   protected Z advanceG(final long k) {
-    return Functions.FACTORIAL.z((int) (k - 1)).negate();
+    return Functions.FACTORIAL.z(k - 1).negate();
   }
 
 }
